+++ pkg/server2/internal/ports/openapi/openapi_non_admin_request_types.gen.go
@@ -39,7 +39,7 @@ type RequestForeignGASPNodeBody struct {
 // RequestSyncResponseBody defines model for RequestSyncResponseBody.
 type RequestSyncResponseBody struct {
 	// Since Timestamp or sequence number from which to start synchronization
-	Since uint32 `json:"since"`
+	Since float64 `json:"since"`
 
 	// Version The version number of the GASP protocol
 	Version int `json:"version"`
